Refactored logging to separate file
