+++ src/graphql/schema.js
@@ -1,29 +1,10 @@
 const { makeExecutableSchema } = require('apollo-server');
-const { applyMiddleware } = require('graphql-middleware');
 const gql = require('graphql-tag');
 const chalk = require('chalk');
 
-const ScalarTypes = require('./scalarTypes');
 const Schemas = require('./app');
-
-// ---
-
-const logTitle = (num) =>
-  console.log(
-    chalk.bold.yellow(
-      `\nCombining ${chalk.bold.gray('(' + num + ')')} app module${num === 1 ? '' : 's'}...`
-    )
-  );
-
-const logListItem = (index, length, name, schema) =>
-  console.log(
-    `  ${index === length - 1 ? '└' : '├'}─ ${chalk.yellow(name)} `.padEnd(42) +
-      chalk.dim.gray('| ') +
-      chalk.cyan('•'.repeat(Object.keys(schema.resolvers?.Query || {}).length)) +
-      chalk.magenta('•'.repeat(Object.keys(schema.resolvers?.Mutation || {}).length))
-  );
-
-// ---
+const ScalarTypes = require('./scalarTypes');
+const { logTitle, logSchemaModuleListItem, applyLoggingToQueryResolvers } = require('./logging');
 
 const rootSchema = {
   _Root: {
@@ -40,35 +21,6 @@ const rootSchema = {
   },
 };
 
-function applyLogging(resolver) {
-  return async (root, args, ctx, info) => {
-    const operationType = info.operation.operation;
-    const operationColor = operationType === 'query' ? chalk.blue : chalk.magenta;
-    const operationName = info.operation.name.value;
-
-    console.log(`\n${operationColor(operationType)} ${chalk.cyan(operationName)}`);
-    console.log(`${operationColor('args')}`);
-    console.log(chalk.gray(JSON.stringify(args, null, 2)));
-
-    return resolver(root, args, ctx, info);
-  };
-}
-
-function applyLoggingToQueryResolvers(resolvers) {
-  if (!resolvers.Query) {
-    return resolvers;
-  }
-
-  return {
-    ...resolvers,
-    Query: Object.fromEntries(
-      Object.entries(resolvers.Query).map(([name, resolverFn]) => {
-        return [name, applyLogging(resolverFn)];
-      })
-    ),
-  };
-}
-
 function combineSchemas({ schemas, options = {} }) {
   logTitle(Object.keys(Schemas).length);
   const schemasFlat = schemas.map((schema) => Object.entries(schema)).flat();
@@ -76,7 +28,7 @@ function combineSchemas({ schemas, options = {} }) {
   let index = 0;
   const combined = schemasFlat.reduce(
     (output, [name, schema]) => {
-      if (schema.appModule) logListItem(index, schemasFlat.length, name, schema);
+      if (schema.appModule) logSchemaModuleListItem(index, schemasFlat.length, name, schema);
       if (schema.typeDefs) output.typeDefs.push(schema.typeDefs);
       if (schema.resolvers)
         output.resolvers.push(
@@ -88,7 +40,7 @@ function combineSchemas({ schemas, options = {} }) {
     { typeDefs: [], resolvers: [] }
   );
 
-  return applyMiddleware(makeExecutableSchema(combined));
+  return makeExecutableSchema(combined);
 }
 
 function makeAppSchema(options) {
